+++ src/MultiTranslator.php
@@ -42,7 +42,7 @@ use yii\log\Logger;
  */
 class MultiTranslator extends Plugin
 {
-    public string $schemaVersion = '1.5.0';
+    public string $schemaVersion = '1.6.0';
     public bool $hasCpSettings = true;
     public bool $hasCpSection = true;
     public ?string $name = 'Multi Translator';
@@ -142,6 +142,7 @@ class MultiTranslator extends Plugin
                 $event->rules['multi-translator/glossaries/edit/<id:\d+>'] = 'multi-translator/glossaries/edit';
                 $event->rules['multi-translator/glossaries/new'] = 'multi-translator/glossaries/new';
                 $event->rules['multi-translator/glossaries/fetch'] = 'multi-translator/glossaries/fetch';
+                $event->rules['multi-translator/style-rules/fetch'] = 'multi-translator/style-rules/fetch';
                 $event->rules['multi-translator/translate/review'] = 'multi-translator/translate/review';
                 $event->rules['multi-translator/field/review'] = 'multi-translator/field/review';
                 $event->rules['multi-translator/field/translate'] = 'multi-translator/field/translate';
@@ -299,6 +300,7 @@ JS;
         $nav = parent::getCpNavItem();
         $nav['subnav']['dashboard'] = ['label' => Craft::t('multi-translator', 'Dashboard'), 'url' => 'multi-translator'];
         $nav['subnav']['glossaries'] = ['label' => Craft::t('multi-translator', 'Glossaries'), 'url' => 'multi-translator/glossaries'];
+        $nav['subnav']['style-rules'] = ['label' => Craft::t('multi-translator', 'Style Rules'), 'url' => 'multi-translator/style-rules'];
 
         if (Craft::$app->user->checkPermission('multiTranslateSettings')) {
             $nav['subnav']['settings'] = ['label' => Craft::t('multi-translator', 'Settings'), 'url' => 'multi-translator/settings'];